Fix occasional jitter
My mistake was assuming that `latest_world_state` was up-to-date,
while it was also the staging buffer for the next frame's world state,
resulting in a torn worldstate whenever entity_refresh messages arrive
before a frame and the corresponding tick_report doesn't arrive until
after the frame.

The fix is to always consume from the interpolation queue, which is
pushed to every frame, and never sees a torn worldstate.